fix(bundle): extract ModelCapabilities type + capabilityField to client-safe files
The previous commit introduced capabilityField() inside cache.ts which
imports @/db. RecoveryPanel (a client component) importing from cache.ts
caused postgres + fs to leak into the browser bundle.

Fix:
- New lib/model-capabilities/types.ts — ModelCapabilities type with no
  server imports (safe for use client components and hooks)
- New lib/model-capabilities/capability-field.ts — capabilityField() with
  no server imports (replaces the unsafe `capability as keyof ModelCapabilities`
  cast; switch statement ensures compile-time exhaustiveness)
- cache.ts re-exports ModelCapabilities from types.ts for backwards
  compat; existing server-only callers are unaffected
- use-model-capabilities.ts, recovery-panel.tsx — import from types.ts /
  capability-field.ts instead of cache.ts
