+++ packages/web/src/lib/model-capabilities/cache.ts
@@ -1,15 +1,9 @@
 import { db } from "@/db";
 import { models } from "@/db/schema";
 import type { ModelCapability } from "@/lib/model-resolver/types";
+import type { ModelCapabilities } from "@/lib/model-capabilities/types";
 
-export type ModelCapabilities = {
-  vision: boolean;
-  documents: boolean;
-  audio: boolean;
-  video: boolean;
-  longContext: boolean;
-  tools: boolean;
-};
+export type { ModelCapabilities } from "@/lib/model-capabilities/types";
 
 let cache: Map<string, ModelCapabilities> | null = null;
 let warnedAboutUnloadedCache = false;
@@ -64,28 +58,6 @@ export function getModelCapabilities(qualifiedModelId: string): ModelCapabilitie
 export function modelHasCapability(qualifiedModelId: string, cap: ModelCapability): boolean {
   const caps = getModelCapabilities(qualifiedModelId);
   if (!caps) return false;
-  switch (cap) {
-    case "vision":
-      return caps.vision;
-    case "documents":
-      return caps.documents;
-    case "audio":
-      return caps.audio;
-    case "video":
-      return caps.video;
-    case "long-context":
-      return caps.longContext;
-    case "tools":
-      return caps.tools;
-  }
-}
-
-/**
- * Returns the `ModelCapabilities` field that corresponds to the given
- * `ModelCapability` string. The mapping is explicit — no string-cast —
- * so adding new capabilities forces a compile-time update here.
- */
-export function capabilityField(caps: ModelCapabilities, cap: ModelCapability): boolean {
   switch (cap) {
     case "vision":
       return caps.vision;
